Extract Navigator from BaseActivity
