+++ app/src/main/java/io/github/droidkaigi/confsched2017/di/ActivityModule.java
@@ -4,25 +4,18 @@ import android.support.v7.app.AppCompatActivity;
 
 import dagger.Module;
 import dagger.Provides;
-import io.github.droidkaigi.confsched2017.view.activity.BaseActivity;
-import io.github.droidkaigi.confsched2017.view.helper.Navigator;
 
 @Module
 public class ActivityModule {
 
-    final BaseActivity activity;
+    final AppCompatActivity activity;
 
-    public ActivityModule(BaseActivity activity) {
+    public ActivityModule(AppCompatActivity activity) {
         this.activity = activity;
     }
 
     @Provides
     public AppCompatActivity activity() {
-        return activity;
-    }
-
-    @Provides
-    public Navigator navigator() {
         return activity;
     }
